+++ packages/smarthr-ui/src/components/FormGroup/FormGroup.tsx
@@ -92,7 +92,7 @@ const classNameGenerator = tv({
 
 const SMARTHR_UI_INPUT_SELECTOR = '[data-smarthr-ui-input="true"]'
 export const CHILDREN_WRAPPER_INPUT_SELECTOR = `.smarthr-ui-FormControl-childrenWrapper ${SMARTHR_UI_INPUT_SELECTOR}`
-const LABEL_TEXT_SELECTOR = '.smarthr-ui-FormControl-labelText'
+export const LABEL_TEXT_SELECTOR = '.smarthr-ui-FormControl-labelText'
 
 export const FormGroup: FC<Props> = ({
   wrapperRef,
@@ -207,64 +207,6 @@ export const FormGroup: FC<Props> = ({
       }
     }
   }, [actualErrorMessages.length, autoBindErrorInput, wrapperRef])
-
-  // HINT: Fieldset内の可視ラベルが無いinputに、legend文言をアクセシブルネームに追加する
-  // https://waic.jp/translations/WCAG21/Understanding/label-in-name.html
-  useEffect(() => {
-    if (!isFieldset || !wrapperRef.current) return
-
-    const labelTextEl = wrapperRef.current.querySelector(LABEL_TEXT_SELECTOR)
-
-    if (!labelTextEl) return
-
-    // HINT: legend変更のたびにaria-labelへ古いlegend文言が蓄積しないよう、
-    // 初回に確定したアクセシブルネームをinput要素ごとに保持しておく
-    const baseAccessibleNames = new WeakMap<HTMLInputElement, string>()
-
-    const updateAriaLabels = () => {
-      const labelText = labelTextEl.textContent || ''
-      if (!labelText) return
-
-      const inputs = wrapperRef.current?.querySelectorAll<HTMLInputElement>(
-        CHILDREN_WRAPPER_INPUT_SELECTOR,
-      )
-      if (!inputs?.length) return
-
-      inputs.forEach((input: HTMLInputElement) => {
-        let accessibleName = baseAccessibleNames.get(input)
-
-        if (accessibleName === undefined) {
-          accessibleName =
-            input.getAttribute('aria-label') ||
-            (input.labels?.[0]?.classList.contains('smarthr-ui-VisuallyHiddenText')
-              ? input.labels[0].textContent || ''
-              : '')
-          baseAccessibleNames.set(input, accessibleName)
-        }
-
-        if (
-          accessibleName &&
-          !accessibleName.includes(labelText) &&
-          !labelText.includes(accessibleName)
-        ) {
-          input.setAttribute('aria-label', `${accessibleName} ${labelText}`)
-        }
-      })
-    }
-
-    // 初回実行
-    updateAriaLabels()
-
-    // label要素の変更を監視
-    const observer = new MutationObserver(updateAriaLabels)
-    observer.observe(labelTextEl, {
-      childList: true,
-      subtree: true,
-      characterData: true,
-    })
-
-    return () => observer.disconnect()
-  }, [isFieldset, wrapperRef])
 
   return (
     <Stack
